datamash: perc: permit non-integer percentiles
This also admits percentiles of 0.0

+++ src/field-ops.h
@@ -88,7 +88,7 @@ struct fieldop
   union {
     long double bin_bucket_size;
     size_t strbin_bucket_size;
-    size_t percentile;
+    long double percentile;
     long double coldness;
     long double trimmed_mean;
     enum extract_number_type get_num_type;
